% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_coda_dic.R
\name{coda.names}
\alias{coda.names}
\title{Get names of node array}
\usage{
coda.names(basename, dim)
}
\arguments{
\item{basename}{the node names}

\item{dim}{dimension of the nodes}
}
\value{
It returns a list of the names of individual elements
}
\description{
This is a helper function from the \code{rjags} library in order to get the names of the individual elements
of a node array. See the package \code{rjags} for more details.
}
\references{
Martyn Plummer (2021). rjags: Bayesian Graphical Models using MCMC. R package version 4-12.
https://CRAN.R-project.org/package=rjags
}
