% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidural_ic-data.R
\docType{data}
\name{epidural_ic}
\alias{epidural_ic}
\title{Meta-analysis data without full compliance information}
\format{
An object of class \code{data.frame} with 27 rows and 14 columns.
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/25592169/}
}
\usage{
data(epidural_ic)
}
\description{
The data contains a meta analysis of the association between using epidural
analgesia in labor and the risk of cesarean section. It contains 27 studies,
only 10 out of which have full compliance information.
}
\examples{
data(epidural_ic)
}
\references{
Bannister-Tyrrell M, Miladinovic B, Roberts CL, Ford JB. 
Adjustment for compliance behavior in trials of epidural analgesia in labor using instrumental variable meta-analysis. 
J Clin Epidemiol. 2015 May;68(5):525-33. doi: 10.1016/j.jclinepi.2014.11.005. Epub 2014 Dec 18. PMID: 25592169.
}
\keyword{dataset}
