% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.meta.c.R
\name{model.meta.c}
\alias{model.meta.c}
\title{Bayesian hierarchical model code for CACE meta-analysis with complete compliance data}
\usage{
model.meta.c(prior.type = "default", Ind = Ind)
}
\arguments{
\item{prior.type}{type of prior information. Default to "default".}

\item{Ind}{specification of individual parameters.}
}
\value{
It returns a model string
}
\description{
This function generates the model code for meta-analysis 
when the dataset has complete compliance information for all studies, 
as described in the paper Section 2.2.2, the Bayesian hierarchical model.
}
\examples{
params.ind <- c(1,1,1,1,1,1,1)
model.string <- model.meta.c(Ind=params.ind)
}
