% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.study.R
\name{model.study}
\alias{model.study}
\title{Model code of CACE analysis for a single study, or a two-step approach for meta-analysis 
with complete complice information}
\usage{
model.study(prior.type = "default", Ind = NULL)
}
\arguments{
\item{prior.type}{type of prior information. Default to "default".}

\item{Ind}{specification of individual parameters.}
}
\value{
It returns a model string
}
\description{
This function generates the model code for a single study using the 
likelihood and model specified in the paper Section 2.1, or a two-step 
approach for meta-analysis with complete compliance information as 
described in the paper Section 2.2 "The two-step approach".
}
\examples{
model.string <- model.study()
}
