% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.cacebayes.R
\name{plot.cacebayes}
\alias{plot.cacebayes}
\title{plot.cacebayes}
\usage{
\method{plot}{cacebayes}(
  obj,
  which = c("trace", "density", "autocorr"),
  param = c("CACE"),
  trialnumber = 1,
  ...
)
}
\arguments{
\item{obj}{an S3 objective to plot}

\item{param}{parameters}

\item{trialnumber}{indicator for the trial number}
}
\value{

}
\description{
plot.cacebayes
}
