% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.forest.R
\name{plot.forest}
\alias{plot.forest}
\title{plot.forest}
\usage{
\method{plot}{forest}(data, obj, ...)
}
\arguments{
\item{data}{data}

\item{obj}{an S3 objective to plot}

\item{...}{others}
}
\value{

}
\description{
plot.forest
}
