% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.noncomp.R
\name{plot.noncomp}
\alias{plot.noncomp}
\title{plot.noncomp}
\usage{
\method{plot}{noncomp}(data, overall = TRUE)
}
\arguments{
\item{data}{data}

\item{overall}{overall}
}
\value{

}
\description{
plot.noncomp
}
