% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt.cacebayes.R
\name{plt.cacebayes}
\alias{plt.cacebayes}
\title{this plot function creates a density plot, a trace plot,
and an acf plot.}
\usage{
plt.cacebayes(
  obj,
  which = c("trace", "density", "autocorr"),
  param = c("CACE"),
  trialnumber = 1
)
}
\arguments{
\item{obj}{an S3 objective to plot}

\item{which}{choose from c("trace", "density", "autocorr")}

\item{param}{parameters}

\item{trialnumber}{indicator for the trial number}
}
\value{
It returns a list of 3 plots.
}
\description{
This function creates several plots of the S3 objective.
}
\examples{
\donttest{
out.meta.c <- cace.meta.c(data = epidural_c, conv.diag = TRUE, 
mcmc.samples = TRUE, study.specific = TRUE)
plt.cacebayes(obj=out.meta.c)
}
}
