% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt.cacebayes.R
\name{plt.density}
\alias{plt.density}
\title{this plot function creates a density plot}
\usage{
plt.density(obj, param = c("CACE"), trialnumber = 1)
}
\arguments{
\item{obj}{an S3 objective to plot}

\item{param}{list of parameters to plot}

\item{trialnumber}{indicator for the trial number}
}
\value{
It returns a density plot in an \code{R} plot window.
}
\description{
This function creates a density plot of the S3 objective.
}
\examples{
\donttest{
out.meta.c <- cace.meta.c(data = epidural_c, conv.diag = TRUE, 
mcmc.samples = TRUE, study.specific = TRUE)
plt.density(obj=out.meta.c)
}
}
