% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt.forest.R
\name{plt.forest}
\alias{plt.forest}
\title{this plot function makes a forest plot.}
\usage{
plt.forest(data, obj, ...)
}
\arguments{
\item{data}{data}

\item{obj}{an S3 objective to plot}

\item{...}{optional parameters passed into the forestplot object}
}
\value{
It returns a forestplot object.
}
\description{
This function provides a visual overview (forst plot).
}
\examples{
\donttest{
data("epidural_c", package = "BayesCACE")
out.meta.c <- cace.meta.c(data = epidural_c, conv.diag = TRUE, 
mcmc.samples = TRUE, study.specific = TRUE)
plt.forest(data=epidural_c, obj=out.meta.c)
}
}
