% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.meta.c.R
\name{prior.meta}
\alias{prior.meta}
\title{The function returns a custom model string specified by the user.}
\usage{
prior.meta(prior.type = "custom")
}
\arguments{
\item{prior.type}{prior type, default is "custom".}
}
\value{
custom model string
}
\description{
The function returns a custom model string specified by the user.
}
\examples{
model.string <- prior.meta()
}
