% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.coda.dic.R
\name{coda.names}
\alias{coda.names}
\title{Get names of node array}
\usage{
coda.names(basename, dim)
}
\arguments{
\item{basename}{the node names}

\item{dim}{dimension of the nodes}
}
\value{
It returns a list of the names of individual elements
}
\description{
This is a helper function from the \code{rjags} library in order to get the names of the individual elements
of a node array. See the package \code{rjags} for more details.
}
\references{
\insertRef{plummer2021rjags}{BayesCACE}
}
