% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.meta.ic.R
\name{model.meta.ic}
\alias{model.meta.ic}
\title{Bayesian hierarchical model code for CACE meta-analysis with complete compliance data}
\usage{
model.meta.ic(prior.type = "default", random.effects = list())
}
\arguments{
\item{prior.type}{type of prior information. Default to "default".}

\item{random.effects}{a list of logical values indicating whether random effects are included in the model.
The list should contain the assignment for these parameters only: \code{delta.n} (\eqn{\delta_{in}}), 
\code{delta.a} (\eqn{\delta_{ia}}), \code{delta.u} (\eqn{\delta_{iu}}), \code{delta.v} (\eqn{\delta_{iv}}), 
\code{delta.s} (\eqn{\delta_{is}}), \code{delta.b} (\eqn{\delta_{ib}}), \code{cor}. The list should be in the
form of \code{list(delta.a = FALSE, cor = FALSE, ...)}. By default, this
is an empty list, and all parameters are default to \code{TRUE}. Parameters that are not listed in the list
are assumed to be \code{TRUE}. Note that \eqn{\rho} (\code{cor}) can only be included when both \eqn{\delta_{in}} 
(\code{delta.n}) and \eqn{\delta_{ia}} (\code{delta.a}) are set to \code{TRUE}. Otherwise, a warning 
occurs and the model continues running by forcing \code{delta.n = TRUE} and \code{delta.a = TRUE}.}
}
\value{
It returns a model string
}
\description{
This function generates the model code for meta-analysis 
when the dataset has incomplete compliance information for all studies, 
as described in the paper Section 2.2.2, the Bayesian hierarchical model.
}
\examples{
# use default settings
model.string <- model.meta.ic()
}
