% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun.coda.dic.R
\name{parse.varname}
\alias{parse.varname}
\title{Parse strings of specific form}
\usage{
parse.varname(varname)
}
\arguments{
\item{varname}{string name of variable}
}
\value{
It returns a list of parsed parameters
}
\description{
This is a helper function from the \code{rjags} library in order to parse the string of form
"a" or "a[n,p:q,r]". See the package \code{rjags} for more details.
}
\references{
\insertRef{plummer2021rjags}{BayesCACE}
}
