% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plt.cacebayes.R
\name{plt.trace}
\alias{plt.trace}
\title{this plot function creates a traceplot}
\usage{
plt.trace(obj, param = c("CACE"), trialnumber = 1, ...)
}
\arguments{
\item{obj}{an S3 objective to plot}

\item{param}{list of parameters to plot}

\item{trialnumber}{indicator for which trial number of the mcmc samples
to use. The default is \code{1}}

\item{...}{optional parameters to pass into the \code{plot} function}
}
\value{
It returns a traceplot in an \code{R} plot window.
}
\description{
This function creates a traceplot of the S3 objective.
}
\examples{
\donttest{
out.meta.c <- cace.meta.c(data = epidural_c, conv.diag = TRUE, 
mcmc.samples = TRUE, study.specific = TRUE)
plt.trace(obj=out.meta.c)
}
}
