% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.study.R
\name{prior.study}
\alias{prior.study}
\title{prior.study}
\usage{
prior.study(prior.type = "custom")
}
\arguments{
\item{prior.type}{prior type, default is "custom"}
}
\value{

}
\description{
prior.study
}
